refactor(levelpacks): move sort to new file, and avoid param re-assign
